+++ spreadsheet-xl/src/main/java/spreadsheet/xlsx/XlsxBookFactory.java
@@ -1,77 +1,55 @@
 /*
- * Copyright 2013 National Bank of Belgium
+ * Copyright 2022 National Bank of Belgium
  *
- * Licensed under the EUPL, Version 1.1 or – as soon they will be approved 
+ * Licensed under the EUPL, Version 1.1 or – as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
  *
  * http://ec.europa.eu/idabc/eupl
  *
- * Unless required by applicable law or agreed to in writing, software 
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package ec.util.spreadsheet.poi;
+package spreadsheet.xlsx;
 
-import ec.util.spreadsheet.helpers.FileHelper;
 import ec.util.spreadsheet.Book;
-import java.io.EOFException;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
+import ec.util.spreadsheet.helpers.FileHelper;
+import nbbrd.service.ServiceProvider;
+import org.checkerframework.checker.nullness.qual.NonNull;
+
+import java.io.*;
 import java.nio.file.AccessDeniedException;
 import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
-import java.util.*;
-import java.util.concurrent.atomic.AtomicBoolean;
-import nbbrd.service.ServiceProvider;
-import org.apache.commons.compress.archivers.zip.Zip64Mode;
-import org.apache.poi.xssf.streaming.SXSSFWorkbook;
-import org.checkerframework.checker.nullness.qual.NonNull;
-import spreadsheet.xlsx.XlsxReader;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 import static java.util.Collections.singletonList;
 
 /**
- *
  * @author Philippe Charles
  */
 @ServiceProvider(Book.Factory.class)
-public class ExcelBookFactory extends Book.Factory {
+public final class XlsxBookFactory extends Book.Factory {
 
-    private static final boolean USE_SHARED_STRINGS = true;
     private static final String XLSX_TYPE = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
     private static final String XLSM_TYPE = "application/vnd.ms-excel.sheet.macroEnabled.12";
 
-    private final AtomicBoolean fast;
-
-    public ExcelBookFactory() {
-        this.fast = new AtomicBoolean(true);
-    }
-
-    //<editor-fold defaultstate="collapsed" desc="Getters/Setters">
-    public void setFast(boolean fast) {
-        this.fast.set(fast);
-    }
-
-    public boolean isFast() {
-        return fast.get();
-    }
-    //</editor-fold>
-
     @Override
     public String getName() {
-        return "Excel";
+        return "Xlsx";
     }
 
     @Override
     public int getRank() {
-        return WRAPPED_RANK;
+        return NATIVE_RANK;
     }
 
     @Override
@@ -96,9 +74,7 @@ public class ExcelBookFactory extends Book.Factory {
     @Override
     public Book load(File file) throws IOException {
         checkFile(file);
-        return fast.get()
-                ? new XlsxReader().read(file.toPath())
-                : PoiBook.create(file);
+        return new XlsxReader().read(file.toPath());
     }
 
     @Override
@@ -106,22 +82,17 @@ public class ExcelBookFactory extends Book.Factory {
         if (stream.available() == 0) {
             throw new EOFException();
         }
-        return fast.get()
-                ? new XlsxReader().read(stream)
-                : PoiBook.create(stream);
+        return new XlsxReader().read(stream);
+    }
+
+    @Override
+    public boolean canStore() {
+        return false;
     }
 
     @Override
     public void store(OutputStream stream, Book book) throws IOException {
-        SXSSFWorkbook target = new SXSSFWorkbook(null, 100, false, USE_SHARED_STRINGS);
-        target.setZip64Mode(Zip64Mode.AsNeeded);
-        try {
-            PoiBookWriter.copy(book, target);
-            target.write(stream);
-        } finally {
-            // dispose of temporary files backing this workbook on disk
-            target.dispose();
-        }
+        throw new UnsupportedOperationException("Not supported");
     }
 
     @NonNull