+++ CHANGELOG.md
@@ -13,11 +13,31 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+## [2.4.0] - 2022-06-14
+
+### Added
+
+- Add `module-info.java` to `spreadsheet-poi`
+- Add media types to identify spreadsheet types
+- Add sorting of book factories by rank
+
+### Changed
+
+- Move Excel native reader to its own module (`spreadsheet-xl`)
+
+### Fixed
+
 ## [2.3.0] - 2022-06-08
 
 ### Added
 
 - Add Maven BOM
+
+### Fixed
+
+- Fix issues with high memory usage (SODS)
+- Fix Zip64 issue (POI)
+- Fix invalid file lock on Windows+JDK8 (POI)
 
 ## [2.2.6] - 2020-03-25
 
@@ -81,7 +101,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - JDK11 cleanup
 
-[Unreleased]: https://github.com/nbbrd/spreadsheet4j/compare/v2.3.0...HEAD
+[Unreleased]: https://github.com/nbbrd/spreadsheet4j/compare/v2.4.0...HEAD
+[2.4.0]: https://github.com/nbbrd/spreadsheet4j/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/nbbrd/spreadsheet4j/compare/v2.2.6...v2.3.0
 [2.2.6]: https://github.com/nbbrd/spreadsheet4j/compare/v2.2.5...v2.2.6
 [2.2.5]: https://github.com/nbbrd/spreadsheet4j/compare/v2.2.4...v2.2.5